Added the update.R function.
This function checks for a newer version of R, and if it exists - it downloads it.

I seperated this function to its own .r file, which also contain a bunch of other functions such as
install.R - for installing R.
check.for.updates.R - for checking if a newer version of R exists.
And turn.version.to.number - which turns a version character (like 2.15.2) to a number which allows  the comparison between versions.