fix no-args bug in `create.global.library`
`create.global.library(global_library_folder)` was failing when passed
no arguments. The expression

```
missing(global_library_folder) | is.null(global_library_folder)
```

was evaluting both sides of the `|` operator - this has been replaced
with a `||` operator (so `is.null(...)` is only evaluated when
`global_library_folder` is not missing).

Tests (and testthat/mockery dependencies) have been added to ensure that
- a user-specified global-library can be used,
- a no-argument call to create.global.library can be made,
- and that a warning is thrown and no library-directory is constructed
  when a non-writeable directory is supplied.

Also,

- rewrote the code that constructs `R_lib_parent` to use
  `dirname(R.home())` for brevity,
- changed the `cat(...)` calls to `message(...)` so that the user can
  suppress them if required.
- add `R.utils` to `DESCRIPTION::Suggests` since required to check Rd
  xrefs during R CMD check
- describe bugfix in NEWS